Fixed errors in and added to certain doc strings.
The tes.V parameter is particularly confusing. This refers to “V” in
the Irwin-Hilton paper (see Figure 3 and Equation 4 and preceding
paragraphs). It is the *Thévenin equivalent voltage* for a circuit
where the actual TES current is supposed to flow through a “load
resistance” (= the shunt resistance plus any parasitic resistance in
the L+TES branch), the inductance, and the TES all in series.

V = I_bias * R_shunt is a constant. This *does not* refer to the
voltage across the TES. So please be careful!