save_positions = (false, false) in pulses callback, plus test for length
Brad was seeing 2 extra points per pulse (e.g. asked for 100 samples, 7
pulses, got 114 samples). This was because the DiscreteCallback was
saving two extra points each time. This was due to an API change,
previously save_everystep needed to be true and something else needed
to be true, now it is independent.